third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-recursive.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#e {
    width: 100px;
    height: 200px;
    background-color: green;
    --foo: {
        @apply --foo;
        background-color: red;
    };
    @apply --foo;
}
</style>
<div id=e>
</div>
<script>
test(function(){
    assert_equals(getComputedStyle(e).backgroundColor, "rgb(0, 128, 0)");
    assert_equals(getComputedStyle(e).getPropertyValue("--foo"), "");
}, "Recursive @apply is invalid");
</script>
